pyproject.toml
[project]
name = "n8n-mcp"
version = "0.1.0"
description = "MCP server + n8n workflow builder and client"
readme = "README.md"
requires-python = ">=3.9,<3.12"
authors = [
  { name = "You", email = "you@example.com" }
]
dependencies = [
  "requests>=2.32.3",
  "pydantic>=2.7.4",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.1",
  "python-dotenv>=1.0.1",
  "gradio>=4.38.1",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.2.2",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "builder",
  "n8n_client",
  "mcp_server",
  "ui_gradio",
]
